fix(testing): moving `mod_backend` and `mod_frontend` to `globals.py` to fix the circular import problem (#23480)
